requirements.txt
streamlit
google-generativeai
Image
python-dotenv
langchain
chromadb
PyPDF2

